+++ app/src/main/java/com/example/comicbookrental/ui/screens/cart/CartViewModel.kt
@@ -50,7 +50,7 @@ class CartViewModel : ViewModel() {
             comicTitle = comic.title,
             comicAuthor = comic.author,
             comicCoverUrl = comic.coverImageUrl,
-            pricePerDay = comic.rentalPrice.toLong(),
+            pricePerDay = comic.rentalPrice,
             startDate = startDate,
             endDate = endDate
         )
@@ -98,13 +98,6 @@ class CartViewModel : ViewModel() {
 
         return rentalDays <= MAX_RENTAL_DAYS
     }
-
-//    private fun String.toPricePerDay(): Long {
-//        // "15,000 ₫" → "15000" → 15000L
-//        return filter { it.isDigit() }
-//            .toLongOrNull()
-//            ?: 0L
-//    }
 
 
 }
